backend/requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-CORS==4.0.0
supabase>=1.0.0
python-dotenv>=1.0.0
groq>=0.9.0